source/ML/requirements.txt
mprpc==0.1.10
numpy==1.10.1
scikit-learn==0.16.1
matplotlib==1.4.3
msgpack-rpc-python==0.4